feat: DELETE /comments 자신이 쓴 댓글만 삭제가능, 자식 댓글이 있을 경우 내용만 삭제 기능 추가

+++ controllers/comments/methods/deleteComment.js
@@ -19,6 +19,16 @@ module.exports = wrapAsync(async (req, res) => {
     return res.status(404).end();
   }
 
+  if (comment.author.toString() !== req.user._id) {
+    res.status(401);
+    return res.json({ message: '자신이 쓴 댓글만 삭제할 수 있습니다.' });
+  }
+
+  if (comment.childComments.length > 0) {
+    await Comment.findByIdAndUpdate(comment._id, { $set: { contents: '삭제된 댓글입니다.', isDeleted: true } });
+    return res.status(204).end();
+  }
+
   await Comment.findByIdAndDelete(id);
 
   return res.status(204).end();
